+++ mobile/app/rules/[game].js
@@ -1,4 +1,4 @@
-import { useLocalSearchParams, useRouter } from 'expo-router';
+bawawdaimport { useLocalSearchParams, useRouter } from 'expo-router';
 import { Image, Pressable, ScrollView, StyleSheet, Text, View } from 'react-native';
 import AppBackground from '../../src/components/shared/AppBackground';
 import {